+++ tsconfig.json
@@ -17,7 +17,7 @@
     "resolveJsonModule": true,
     "jsx": "preserve",
     "incremental": true,
-    "isolatedModules": false
+    "isolatedModules": true
   },
   "include": [
     "next-env.d.ts",
